requirements.txt
fastapi==0.85.0
uvicorn==0.19.0
python-dotenv==0.21.0